+++ .github/workflows/build-energy-origin-event-store.yaml
@@ -1,0 +1,54 @@
+name: Build - EnergyOriginEventStore
+
+on:
+  pull_request:
+    paths:
+      - libraries/dotnet/EnergyOriginEventStore/**
+  push:
+    branches:
+      - main
+    paths:
+      - libraries/dotnet/EnergyOriginEventStore/**
+
+concurrency:
+  group: ${{ github.ref }}-EnergyOriginEventStore
+  cancel-in-progress: false
+
+defaults:
+  run:
+    working-directory: ./libraries/dotnet/EnergyOriginEventStore
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+
+    steps:
+    - uses: actions/checkout@v3
+
+    - name: Get version
+      id: get_version
+      uses: Energinet-Datahub/.github/.github/actions/yaml-get@main
+      with:
+        yaml_file: ./libraries/dotnet/EnergyOriginEventStore/version.yaml
+        yaml_path: version
+
+    - name: Setup .NET
+      uses: actions/setup-dotnet@v1.9.0
+      with:
+        dotnet-version: 6.0.101
+
+    - name: Restore
+      run: dotnet restore
+
+    - name: Lint
+      run: dotnet format --verify-no-changes -v diag
+
+    - name: Test
+      run: dotnet test --no-restore --verbosity normal *.Tests
+
+    - name: Pack
+      run: dotnet pack --output build --configuration Release --no-restore -p:Version=${{ steps.get_version.outputs.result }} EnergyOriginEventStore
+
+    - name: Publish
+      if: startsWith(github.ref, 'refs/heads/main')
+      run: dotnet nuget push build/*.nupkg --api-key ${{ secrets.NUGET_API_KEY }} --source https://api.nuget.org/v3/index.json
