--- domains/event-store/.devcontainer/devcontainer.json
@@ -1,8 +1,0 @@
-
-{	"name": "Debian",
-	"build": {
-		"dockerfile": "Dockerfile",
-		"args": { "VARIANT": "bullseye" }
-	},
-	"remoteUser": "vscode"
-}
