--- domains/event-store/.gitignore
@@ -1,13 +1,0 @@
-# Created by https://www.toptal.com/developers/gitignore/api/dotnetcore
-# Edit at https://www.toptal.com/developers/gitignore?templates=dotnetcore
-
-### DotnetCore ###
-# .NET Core build folders
-bin/
-obj/
-
-# Common node modules locations
-/node_modules
-/wwwroot/node_modules
-
-# End of https://www.toptal.com/developers/gitignore/api/dotnetcore
