+++ libraries/dotnet/EnergyOriginEventStore/EnergyOriginEventStore.sln
@@ -1,0 +1,28 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30114.105
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EnergyOriginEventStore", "EnergyOriginEventStore\EnergyOriginEventStore.csproj", "{D250E765-78FF-4E7C-B66C-51395EED9014}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EnergyOriginEventStore.Tests", "EnergyOriginEventStore.Tests\EnergyOriginEventStore.Tests.csproj", "{79698CB6-956A-4C26-8974-FFE256812945}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{D250E765-78FF-4E7C-B66C-51395EED9014}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D250E765-78FF-4E7C-B66C-51395EED9014}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D250E765-78FF-4E7C-B66C-51395EED9014}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D250E765-78FF-4E7C-B66C-51395EED9014}.Release|Any CPU.Build.0 = Release|Any CPU
+		{79698CB6-956A-4C26-8974-FFE256812945}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{79698CB6-956A-4C26-8974-FFE256812945}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{79698CB6-956A-4C26-8974-FFE256812945}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{79698CB6-956A-4C26-8974-FFE256812945}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
