+++ libraries/dotnet/EnergyOriginEventStore/EnergyOriginEventStore/EnergyOriginEventStore.csproj
@@ -1,0 +1,21 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net6.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+    <IsPackable>true</IsPackable>
+    <PackageId>EnergyOriginEventStore</PackageId>
+    <RepositoryUrl>https://github.com/Energinet-DataHub/energy-origin</RepositoryUrl>
+    <Authors>Team Fenris</Authors>
+    <Company>Energinet DataHub A/S</Company>
+    <PackageDescription>This library contains EnergyOrigins EventStore interfaces and implementations</PackageDescription>
+    <PackageLicenseExpression>Apache-2.0</PackageLicenseExpression>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
+    <PackageReference Include="EnergyOriginDateTimeExtension" Version="1.1.3" />
+  </ItemGroup>
+
+</Project>
