+++ libraries/dotnet/EnergyOriginEventStore/EnergyOriginEventStore/EventStore/EventStoreFactory.cs
