+++ libraries/dotnet/EnergyOriginEventStore/EnergyOriginEventStore/EventStore/FlatFile/FlatFileEventConsumer.cs
