+++ libraries/dotnet/EnergyOriginEventStore/EnergyOriginEventStore/EventStore/FlatFile/FlatFileEventStore.cs
