+++ libraries/dotnet/EnergyOriginEventStore/EnergyOriginEventStore/EventStore/IEventConsumer.cs
