+++ libraries/dotnet/EnergyOriginEventStore/EnergyOriginEventStore/EventStore/IEventStore.cs
