+++ libraries/dotnet/EnergyOriginEventStore/EnergyOriginEventStore/EventStore/Serialization/Event.cs
