+++ libraries/dotnet/EnergyOriginEventStore/EnergyOriginEventStore/EventStore/Serialization/EventModel.cs
