+++ libraries/dotnet/EnergyOriginEventStore/EnergyOriginEventStore/Topics/Said.cs
