+++ libraries/dotnet/EnergyOriginEventStore/EnergyOriginEventStore/Topics/Unpack.cs
