+++ libraries/dotnet/EnergyOriginEventStore/EnergyOriginEventStore/Topics/UserCreated.cs
