+++ libraries/dotnet/EnergyOriginEventStore/EnergyOriginEventStore/Topics/UserCreatedVersion1.cs
